+++ src/account/DepartmentAccount.java
@@ -4,7 +4,7 @@ import person.Person;
 
 public class DepartmentAccount extends Account {
 
-    public DepartmentAccount(long Id, String password, Person person, Position position) {
+    public DepartmentAccount(int Id, String password, Person person, Position position) {
         super(Id, password, person, position);
     }
